+++ src/main/java/com/atlassian/db/replica/internal/ReplicaCallableStatement.java
@@ -40,7 +40,8 @@ public class ReplicaCallableStatement extends ReplicaPreparedStatement implement
         Integer resultSetConcurrency,
         Integer resultSetHoldability,
         Set<String> readOnlyFunctions,
-        DualConnection dualConnection
+        DualConnection dualConnection,
+        boolean compatibleWithPreviousVersion
     ) {
         super(
             connectionProvider,
@@ -51,7 +52,8 @@ public class ReplicaCallableStatement extends ReplicaPreparedStatement implement
             resultSetConcurrency,
             resultSetHoldability,
             readOnlyFunctions,
-            dualConnection
+            dualConnection,
+            compatibleWithPreviousVersion
         );
         this.sql = sql;
         this.resultSetType = resultSetType;
@@ -641,6 +643,7 @@ public class ReplicaCallableStatement extends ReplicaPreparedStatement implement
         private final String sql;
         private final Set<String> readOnlyFunctions;
         private final DualConnection dualConnection;
+        private final boolean compatibleWithPreviousVersion;
         private Integer resultSetType;
         private Integer resultSetConcurrency;
         private Integer resultSetHoldability;
@@ -651,7 +654,8 @@ public class ReplicaCallableStatement extends ReplicaPreparedStatement implement
             DatabaseCall databaseCall,
             String sql,
             Set<String> readOnlyFunctions,
-            DualConnection dualConnection
+            DualConnection dualConnection,
+            boolean compatibleWithPreviousVersion
         ) {
             this.connectionProvider = connectionProvider;
             this.consistency = consistency;
@@ -659,6 +663,7 @@ public class ReplicaCallableStatement extends ReplicaPreparedStatement implement
             this.sql = sql;
             this.readOnlyFunctions = readOnlyFunctions;
             this.dualConnection = dualConnection;
+            this.compatibleWithPreviousVersion = compatibleWithPreviousVersion;
         }
 
         public ReplicaCallableStatement.Builder resultSetType(int resultSetType) {
@@ -686,7 +691,8 @@ public class ReplicaCallableStatement extends ReplicaPreparedStatement implement
                 resultSetConcurrency,
                 resultSetHoldability,
                 readOnlyFunctions,
-                dualConnection
+                dualConnection,
+                compatibleWithPreviousVersion
             );
         }
     }