+++ src/main/java/com/atlassian/db/replica/internal/ReplicaStatement.java
@@ -41,7 +41,7 @@ public class ReplicaStatement implements Statement {
     private final DecisionAwareReference<Statement> readStatement;
     private final DecisionAwareReference<Statement> writeStatement;
     private final DualConnection dualConnection;
-
+    private final boolean compatibleWithPreviousVersion;
     public ReplicaStatement(
         ReplicaConsistency consistency,
         ReplicaConnectionProvider connectionProvider,
@@ -50,7 +50,8 @@ public class ReplicaStatement implements Statement {
         Integer resultSetConcurrency,
         Integer resultSetHoldability,
         Set<String> readOnlyFunctions,
-        DualConnection dualConnection
+        DualConnection dualConnection,
+        boolean compatibleWithPreviousVersion
     ) {
         this.consistency = consistency;
         this.connectionProvider = connectionProvider;
@@ -72,6 +73,7 @@ public class ReplicaStatement implements Statement {
             }
         };
         this.dualConnection = dualConnection;
+        this.compatibleWithPreviousVersion = compatibleWithPreviousVersion;
     }
 
     @Override
@@ -193,7 +195,7 @@ public class ReplicaStatement implements Statement {
         checkClosed();
         final RouteDecisionBuilder decisionBuilder;
         final Statement statement;
-        SqlQuery sqlQuery = new SqlQuery(sql);
+        SqlQuery sqlQuery = new SqlQuery(sql, compatibleWithPreviousVersion);
         if (sqlQuery.isSqlSet()) {
             decisionBuilder = new RouteDecisionBuilder(READ_OPERATION).sql(sql);
             statement = getReadStatement(decisionBuilder);
@@ -560,14 +562,16 @@ public class ReplicaStatement implements Statement {
         ReplicaConsistency consistency,
         DatabaseCall databaseCall,
         Set<String> readOnlyFunctions,
-        DualConnection dualConnection
+        DualConnection dualConnection,
+        boolean compatibleWithPreviousVersion
     ) {
         return new Builder(
             connectionProvider,
             consistency,
             databaseCall,
             readOnlyFunctions,
-            dualConnection
+            dualConnection,
+            compatibleWithPreviousVersion
         );
     }
 
@@ -588,7 +592,7 @@ public class ReplicaStatement implements Statement {
         }
         String sql = decisionBuilder.getSql();
         if (sql != null) {
-            SqlQuery sqlQuery = new SqlQuery(sql);
+            SqlQuery sqlQuery = new SqlQuery(sql, compatibleWithPreviousVersion);
             if (sqlQuery.isWriteOperation(sqlFunction)) {
                 decisionBuilder.reason(WRITE_OPERATION);
                 return prepareWriteStatement(decisionBuilder);
@@ -640,6 +644,7 @@ public class ReplicaStatement implements Statement {
         private final DatabaseCall databaseCall;
         private final Set<String> readOnlyFunctions;
         private final DualConnection dualConnection;
+        private final boolean compatibleWithPreviousVersion;
         private Integer resultSetType;
         private Integer resultSetConcurrency;
         private Integer resultSetHoldability;
@@ -649,13 +654,15 @@ public class ReplicaStatement implements Statement {
             ReplicaConsistency consistency,
             DatabaseCall databaseCall,
             Set<String> readOnlyFunctions,
-            DualConnection dualConnection
+            DualConnection dualConnection,
+            boolean compatibleWithPreviousVersion
         ) {
             this.connectionProvider = connectionProvider;
             this.consistency = consistency;
             this.databaseCall = databaseCall;
             this.readOnlyFunctions = readOnlyFunctions;
             this.dualConnection = dualConnection;
+            this.compatibleWithPreviousVersion = compatibleWithPreviousVersion;
         }
 
         public Builder resultSetType(int resultSetType) {
@@ -682,7 +689,8 @@ public class ReplicaStatement implements Statement {
                 resultSetConcurrency,
                 resultSetHoldability,
                 readOnlyFunctions,
-                dualConnection
+                dualConnection,
+                compatibleWithPreviousVersion
             );
         }
     }