+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.7.0...master
 
+### Fixed
+- Set connection runtime parameters, after switching to the main connection (cover uppercase SET)
+
 ## [2.7.0] - 2022-09-08
 [2.7.0]: https://github.com/atlassian-labs/db-replica/compare/release-2.6.4...release-2.7.0